Improve backtrace when key doesn't exist
Previously, the backtrace would include the `method_missing` method,
which was confusing since the implementation details of the nested chest
item would not be obvious to a consumer.

Using `caller` for the backtrace of the raised error reports the exact
line that key access was made one, rather than the line within
`method_missing`, which makes it easier to read.