Allow key to be unlocked via environment variable in API
This is useful in CI contexts where we don't want to have to deal with
the unprotected key file.

This required a refactor to instantiate chests with a key directly
instead of the key file path, since we don't want to require the
existence of an unlocked key file in all cases.