Remove ability to access chest keys via method calls
Our method_missing implementation proved problematic because `Kernel`
defines a `test` method, so when trying to access a key named test (as
in the Rails test environment) it would attempt to call `Kernel#test`.

To prevent confusion, we're going to remove support for this notation,
and change the hash access implementation to raise an error if the key
doesn't exist.