Add ability to access chest keys via method calls
Our removal of this in 1bfc105 was unfortunately premature. While
there's no way to avoid the scenario where the incorrect `test` method
is invoked, we need to define this in order to work nicely with Rails,
since it expects things like `secret_key_base` to be accessible via
method call.

The documentation will still encourage accessing via hash, however.