Don't delete `ARCANUS_PASSWORD` environment variable
We originally deleted this variable after use in order to prevent the
password accidentally being leaked (since it's stored in the process
environment, its accessible to everything in that process space).

However, this breaks tests run via `parallel_tests`, since that forks
and each fork needs to unlock its Arcanus chest again (why this is the
case is unclear, since ideally the child shares the same memory as
the parent).

In the interest of getting this working with `parallel_tests`, however,
remove the deletion.