Use ARCANUS_PASSWORD when running arcanus CLI commands
Allowing us to run `arcanus shell --export` without having previously
run `arcanus unlock` is potentially useful in CI environments.

Add support for this in the unlock code. Note that this is potentially
undesirable in that we permanently unlock the chest by saving the
unprotected key file, but to not do this would require some refactoring
that I lack the spare time for at the moment.