Allow encryption of values of any size
We previously had a limitation on the size of the item we could encrypt
due to the algorithm we were using. Switch to a block mode cipher using
a randomly generated symmetric key (which itself is encrypted using the
RSA key).

This is a breaking change, but the value is worth it and the upgrade
path is pretty easy.