Remove unused childprocess dependency
It's not clear why this was ever added. Perhaps we thought we would need
to invoke an external process while capturing output.

However, since that never came to pass, it seems silly to include it as
a dependency. Remove it since we never utilize the `spawn` helper
anywhere in the codebase.